refactor: move stops to core
