+++ packages/core/src/index.ts
@@ -1,8 +1,12 @@
-export type * from './registries/ContextMenuRegistry';
-export type * from './registries/SlashOptionRegistry';
+export type * from '@registries/ContextMenuRegistry';
+export type * from '@registries/SlashOptionRegistry';
 
-export { BuilderComponent, RowComponent } from './components/Component';
-export { type RowType, type BuilderType } from './components/builderTypes';
+export { BuilderComponent, RowComponent } from '@components/Component';
+export { type RowType, type BuilderType } from '@components/builderTypes';
+
+export { Notice } from '@stops/Notice';
+export { Fault } from '@stops/Fault';
+export { Silence } from '@stops/Silence';
 
 /** Package version */
 export const version = process.env.PACKAGE_VERSION ?? '0.0.0';
