+++ packages/core/src/stops/NoticeCard.ts
@@ -1,5 +1,6 @@
 import { TextDisplayBuilder } from '@discordjs/builders';
-import { BuilderComponent } from '@seedcord/core';
+
+import { BuilderComponent } from '@components/Component';
 
 /**
  * Built fresh inside a {@link Notice}'s `render` to back its ComponentsV2 reply. The title renders as
