+++ packages/core/src/stops/Silence.ts
@@ -7,7 +7,7 @@
  *
  * @example
  * ```ts
- * import { Silence } from '@seedcord/kit';
+ * import { Silence } from 'seedcord';
  *
  * // before any reply or defer, drop the interaction with no reply and no report
  * if (await isBlacklisted(interaction.user.id)) throw new Silence('blacklisted user');
