+++ packages/core/tests/stops/NoticeCard.test.ts
@@ -1,8 +1,8 @@
 import { ContainerBuilder } from '@discordjs/builders';
-import { setBotColor } from '@seedcord/core/internal';
-import { ComponentType } from 'discord.js';
+import { ComponentType } from 'discord-api-types/v10';
 import { afterEach, describe, expect, it } from 'vitest';
 
+import { setBotColor } from '@components/botColorHolder';
 import { NoticeCard } from '@stops/NoticeCard';
 
 function cardText(card: ContainerBuilder): string {
