+++ packages/core/tests/stops/Silence.test.ts
