+++ packages/core/tsconfig.json
@@ -3,8 +3,14 @@
     "extends": "@seedcord/tsconfig/node",
     "compilerOptions": {
         "rootDir": "./",
-        "outDir": "./dist"
+        "outDir": "./dist",
+        "paths": {
+            "@src/*": ["./src/*"],
+            "@components/*": ["./src/components/*"],
+            "@registries/*": ["./src/registries/*"],
+            "@stops/*": ["./src/stops/*"]
+        }
     },
-    "include": ["src/**/*.ts", "tests/**/*.ts", "tsdown.config.ts"],
+    "include": ["src/**/*.ts", "tests/**/*.ts", "tsdown.config.ts", "vitest.config.ts"],
     "exclude": ["node_modules", "dist"]
 }
