+++ packages/kit/src/index.ts
@@ -1,6 +1,3 @@
-export { Notice } from '@stops/Notice';
-export { Fault } from '@stops/Fault';
-export { Silence } from '@stops/Silence';
 export { CustomId } from '@customId/CustomId';
 export { paginate } from '@pagination/paginate';
 export { type PageView } from '@pagination/PageView';
