+++ packages/kit/src/internal.index.ts
@@ -2,5 +2,3 @@ export { prefixOf, decodeFor, type AnyCustomId } from '@customId/CustomId';
 export { PAGE_MAX, pageCursor, type PageCursor } from '@pagination/cursor';
 export { ComponentDefsKey, type HasComponentDefs } from '@customId/routing';
 export type { DecodedParams } from '@customId/Field';
-
-export { NoticeCard } from '@stops/NoticeCard';
