+++ packages/kit/tsconfig.json
@@ -7,8 +7,7 @@
         "paths": {
             "@src/*": ["./src/*"],
             "@customId/*": ["./src/customId/*"],
-            "@pagination/*": ["./src/pagination/*"],
-            "@stops/*": ["./src/stops/*"]
+            "@pagination/*": ["./src/pagination/*"]
         },
         "types": []
     },
