+++ packages/seedcord/src/bot/gates/catalog/IgnoreBots.ts
@@ -1,4 +1,4 @@
-import { Silence } from '@seedcord/kit';
+import { Silence } from '@seedcord/core';
 
 import { defineGate } from '../Gate';
