+++ packages/seedcord/tests/customId/handlers.test.ts
@@ -1,4 +1,5 @@
-import { CustomId, Notice } from '@seedcord/kit';
+import { Notice } from '@seedcord/core';
+import { CustomId } from '@seedcord/kit';
 import { describe, expect, it } from 'vitest';
 
 import { ButtonRoute, ModalRoute, SelectMenuRoute, SelectMenuKind } from '@bDecorators/Interactions';
