+++ packages/seedcord/tests/utils/TestNotice.ts
@@ -1,4 +1,4 @@
-import { Notice } from '@seedcord/kit';
+import { Notice } from '@seedcord/core';
 
 import type { ReplyResponse } from '@seedcord/types';
